Move token specific functions to separate files
